made transformer model serializable, rearrange scripts
